Complete user login/registration, Modals and update README.md
Ammended the existing code and completed the user registration and
login code for the interface, although not perfect. It works. Also
Also, I made a few changes to README.md to better reflect the background
of the project and also a few random fixes here and there.
I've also included a modal for the Login form, and a simple script
"functions.php" that takes in an action and values and does it's magic.
I'll be adding more to it later. Too sleepy to continue, especially
since it's 5.30 AM.